requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-extensions==3.2.3
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
psycopg2-binary==2.9.10
PyJWT==2.10.0
PyPDF2==3.0.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
